refactor AjaxResult object to R1

+++ src/main/java/com/mercury/mallproject/common/enumresource/ResultCodeEnum.java
@@ -3,7 +3,7 @@ package com.mercury.mallproject.common.enumresource;
 /**
  * 常用API操作码
  */
-public enum ResultCodeEnum implements EnumCode<Long> {
+public enum ResultCodeEnum implements EnumCode<Integer> {
     SUCCESS(0, "操作成功"),
     FAILED(1, "操作失败"),
     VALIDATE_FAILED(404, "参数检验失败"),
@@ -13,17 +13,17 @@ public enum ResultCodeEnum implements EnumCode<Long> {
     CAPTCHA_ERROR(500, "验证码错误"),
     PASSWORD_ERROR(500, "登陆密码错误"),
     USER_NOT_FOUNT(500, "用户不存在");
-    private long key;
+    private int key;
     private String description;
 
-    private ResultCodeEnum(long code, String message) {
+    private ResultCodeEnum(int code, String message) {
         this.key = code;
         this.description = message;
     }
 
 
     @Override
-    public Long getCode() {
+    public Integer getCode() {
         return key;
     }
